+++ src/main/types.ts
@@ -2,3 +2,4 @@ import type { Core } from '../core/core.js';
 
 
 export type FlavorFunction = (common: Core) => (void | Promise<void>);
+export type PackageManager = 'npm' | 'yarn' | 'pnpm';